Separate test case to `AnnotatedHttpServiceBlockingTest`
