+++ core/src/test/java/com/linecorp/armeria/internal/annotation/AnnotatedHttpServiceTest.java
@@ -27,14 +27,8 @@ import java.util.List;
 import java.util.Optional;
 import java.util.Set;
 import java.util.TreeSet;
-import java.util.concurrent.BlockingQueue;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CompletionStage;
-import java.util.concurrent.LinkedTransferQueue;
-import java.util.concurrent.ThreadFactory;
-import java.util.concurrent.ThreadPoolExecutor;
-import java.util.concurrent.TimeUnit;
-import java.util.concurrent.atomic.AtomicInteger;
 import java.util.stream.Collectors;
 
 import javax.annotation.Nullable;
@@ -50,7 +44,6 @@ import org.apache.http.impl.client.HttpClients;
 import org.apache.http.message.BasicNameValuePair;
 import org.apache.http.protocol.HTTP;
 import org.apache.http.util.EntityUtils;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
 
@@ -74,7 +67,6 @@ import com.linecorp.armeria.common.Request;
 import com.linecorp.armeria.common.RequestContext;
 import com.linecorp.armeria.common.RequestHeaders;
 import com.linecorp.armeria.common.ResponseHeaders;
-import com.linecorp.armeria.common.util.ThreadFactories;
 import com.linecorp.armeria.server.HttpStatusException;
 import com.linecorp.armeria.server.ServerBuilder;
 import com.linecorp.armeria.server.ServiceRequestContext;
@@ -99,27 +91,7 @@ import com.linecorp.armeria.server.logging.LoggingService;
 import com.linecorp.armeria.testing.internal.AnticipatedException;
 import com.linecorp.armeria.testing.junit.server.ServerExtension;
 
-public class AnnotatedHttpServiceTest {
-
-    private static final CountingThreadPoolExecutor executor = new CountingThreadPoolExecutor(
-            0, 1, 1, TimeUnit.SECONDS, new LinkedTransferQueue<>(),
-            ThreadFactories.newThreadFactory("blocking-test", true));
-
-    private static final AtomicInteger blockingCount = new AtomicInteger();
-
-    private static class CountingThreadPoolExecutor extends ThreadPoolExecutor {
-
-        CountingThreadPoolExecutor(int corePoolSize, int maximumPoolSize, long keepAliveTime,
-                                   TimeUnit unit, BlockingQueue<Runnable> workQueue,
-                                   ThreadFactory threadFactory) {
-            super(corePoolSize, maximumPoolSize, keepAliveTime, unit, workQueue, threadFactory);
-        }
-
-        @Override
-        protected void beforeExecute(Thread t, Runnable r) {
-            blockingCount.incrementAndGet();
-        }
-    }
+class AnnotatedHttpServiceTest {
 
     @RegisterExtension
     static final ServerExtension server = new ServerExtension() {
@@ -162,11 +134,6 @@ public class AnnotatedHttpServiceTest {
 
             sb.annotatedService("/11", new MyAnnotatedService11(),
                                 LoggingService.newDecorator());
-            sb.annotatedService("/12", new MyAnnotatedService12(),
-                                LoggingService.newDecorator());
-            sb.annotatedService("/13", new MyAnnotatedService13(),
-                                LoggingService.newDecorator());
-            sb.blockingTaskExecutor(executor, false);
         }
     };
 
@@ -766,11 +733,6 @@ public class AnnotatedHttpServiceTest {
         }
     }
 
-    @BeforeEach
-    void clear() {
-        blockingCount.set(0);
-    }
-
     @Test
     public void testAnnotatedHttpService() throws Exception {
         try (CloseableHttpClient hc = HttpClients.createMinimal()) {
@@ -1082,64 +1044,6 @@ public class AnnotatedHttpServiceTest {
             testStatusCode(hc, get("/1/void/204"), 204);
             testBodyAndContentType(hc, get("/1/void/200"), "200 OK", MediaType.PLAIN_TEXT_UTF_8.toString());
         }
-    }
-
-    @Test
-    public void testOriginBlockingTaskType() throws Exception {
-        final HttpClient client = HttpClient.of(server.uri("/"));
-
-        String path = "/12/httpResponse";
-        RequestHeaders headers = RequestHeaders.of(HttpMethod.GET, path);
-        AggregatedHttpResponse res = client.execute(headers).aggregate().join();
-        assertThat(res.status()).isSameAs(HttpStatus.OK);
-        assertThat(blockingCount).hasValue(0);
-
-        path = "/12/aggregatedHttpResponse";
-        headers = RequestHeaders.of(HttpMethod.GET, path);
-        res = client.execute(headers).aggregate().join();
-        assertThat(res.status()).isSameAs(HttpStatus.OK);
-        assertThat(blockingCount).hasValue(0);
-
-        path = "/12/jsonNode";
-        headers = RequestHeaders.of(HttpMethod.GET, path);
-        res = client.execute(headers).aggregate().join();
-        assertThat(res.status()).isSameAs(HttpStatus.OK);
-        assertThat(blockingCount).hasValue(0);
-
-        path = "/12/completionStage";
-        headers = RequestHeaders.of(HttpMethod.GET, path);
-        res = client.execute(headers).aggregate().join();
-        assertThat(res.status()).isSameAs(HttpStatus.OK);
-        assertThat(blockingCount).hasValue(0);
-    }
-
-    @Test
-    public void testOnlyBlockingTaskType() throws Exception {
-        final HttpClient client = HttpClient.of(server.uri("/"));
-
-        String path = "/13/httpResponse";
-        RequestHeaders headers = RequestHeaders.of(HttpMethod.GET, path);
-        AggregatedHttpResponse res = client.execute(headers).aggregate().join();
-        assertThat(res.status()).isSameAs(HttpStatus.OK);
-        assertThat(blockingCount).hasValue(1);
-
-        path = "/13/aggregatedHttpResponse";
-        headers = RequestHeaders.of(HttpMethod.GET, path);
-        res = client.execute(headers).aggregate().join();
-        assertThat(res.status()).isSameAs(HttpStatus.OK);
-        assertThat(blockingCount).hasValue(2);
-
-        path = "/13/jsonNode";
-        headers = RequestHeaders.of(HttpMethod.GET, path);
-        res = client.execute(headers).aggregate().join();
-        assertThat(res.status()).isSameAs(HttpStatus.OK);
-        assertThat(blockingCount).hasValue(3);
-
-        path = "/13/completionStage";
-        headers = RequestHeaders.of(HttpMethod.GET, path);
-        res = client.execute(headers).aggregate().join();
-        assertThat(res.status()).isSameAs(HttpStatus.OK);
-        assertThat(blockingCount).hasValue(4);
     }
 
     private enum UserLevel {
